Working on issue #4. This commit includes bug fixes which should be a step toward saving the currency values with the correct decimal places.

+++ ShoppingBirdData/Stored Procedures/usp_InsertItemPrice.sql
@@ -7,7 +7,7 @@ CREATE PROCEDURE [dbo].[usp_InsertItemPrice]
 	@Barcode VARCHAR(50),
     @TaxId INT , 
     @StoreId INT , 
-    @RetailPrice INT, 
+    @RetailPrice DECIMAL(13,4), 
     @UnitId INT, 
     @UpdatedAt DATE, 
     @CreatedAt DATE
